Extract HTTP Proxy tests from test_open-uri.rb
