+++ test/open-uri/test_open-uri.rb
@@ -267,170 +267,6 @@ class TestOpenURI < Test::Unit::TestCase
     }
   end
 
-  def test_proxy
-    with_http {|srv, dr, url|
-      proxy_log = StringIO.new(''.dup)
-      proxy_logger = WEBrick::Log.new(proxy_log, WEBrick::BasicLog::WARN)
-      proxy_auth_log = ''.dup
-      proxy = WEBrick::HTTPProxyServer.new({
-        :ServerType => Thread,
-        :Logger => proxy_logger,
-        :AccessLog => [[NullLog, ""]],
-        :ProxyAuthProc => lambda {|req, res|
-          proxy_auth_log << req.request_line
-        },
-        :BindAddress => '127.0.0.1',
-        :Port => 0})
-      _, proxy_port, _, proxy_host = proxy.listeners[0].addr
-      proxy_url = "http://#{proxy_host}:#{proxy_port}/"
-      begin
-        proxy_thread = proxy.start
-        srv.mount_proc("/proxy", lambda { |req, res| res.body = "proxy" } )
-        URI.open("#{url}/proxy", :proxy=>proxy_url) {|f|
-          assert_equal("200", f.status[0])
-          assert_equal("proxy", f.read)
-        }
-        assert_match(/#{Regexp.quote url}/, proxy_auth_log); proxy_auth_log.clear
-        URI.open("#{url}/proxy", :proxy=>URI(proxy_url)) {|f|
-          assert_equal("200", f.status[0])
-          assert_equal("proxy", f.read)
-        }
-        assert_match(/#{Regexp.quote url}/, proxy_auth_log); proxy_auth_log.clear
-        URI.open("#{url}/proxy", :proxy=>nil) {|f|
-          assert_equal("200", f.status[0])
-          assert_equal("proxy", f.read)
-        }
-        assert_equal("", proxy_auth_log); proxy_auth_log.clear
-        assert_raise(ArgumentError) {
-          URI.open("#{url}/proxy", :proxy=>:invalid) {}
-        }
-        assert_equal("", proxy_auth_log); proxy_auth_log.clear
-        with_env("http_proxy"=>proxy_url) {
-          # should not use proxy for 127.0.0.0/8.
-          URI.open("#{url}/proxy") {|f|
-            assert_equal("200", f.status[0])
-            assert_equal("proxy", f.read)
-          }
-        }
-        assert_equal("", proxy_auth_log); proxy_auth_log.clear
-      ensure
-        proxy.shutdown
-        proxy_thread.join
-      end
-      assert_equal("", proxy_log.string)
-    }
-  end
-
-  def test_proxy_http_basic_authentication_failure
-    with_http {|srv, dr, url|
-      proxy_log = StringIO.new(''.dup)
-      proxy_logger = WEBrick::Log.new(proxy_log, WEBrick::BasicLog::WARN)
-      proxy_auth_log = ''.dup
-      proxy = WEBrick::HTTPProxyServer.new({
-        :ServerType => Thread,
-        :Logger => proxy_logger,
-        :AccessLog => [[NullLog, ""]],
-        :ProxyAuthProc => lambda {|req, res|
-          proxy_auth_log << req.request_line
-          if req["Proxy-Authorization"] != "Basic #{['user:pass'].pack('m').chomp}"
-            raise WEBrick::HTTPStatus::ProxyAuthenticationRequired
-          end
-        },
-        :BindAddress => '127.0.0.1',
-        :Port => 0})
-      _, proxy_port, _, proxy_host = proxy.listeners[0].addr
-      proxy_url = "http://#{proxy_host}:#{proxy_port}/"
-      begin
-        th = proxy.start
-        srv.mount_proc("/proxy", lambda { |req, res| res.body = "proxy" } )
-        exc = assert_raise(OpenURI::HTTPError) { URI.open("#{url}/proxy", :proxy=>proxy_url) {} }
-        assert_equal("407", exc.io.status[0])
-        assert_match(/#{Regexp.quote url}/, proxy_auth_log); proxy_auth_log.clear
-      ensure
-        proxy.shutdown
-        th.join
-      end
-      assert_match(/ERROR WEBrick::HTTPStatus::ProxyAuthenticationRequired/, proxy_log.string)
-    }
-  end
-
-  def test_proxy_http_basic_authentication_success
-    with_http {|srv, dr, url|
-      proxy_log = StringIO.new(''.dup)
-      proxy_logger = WEBrick::Log.new(proxy_log, WEBrick::BasicLog::WARN)
-      proxy_auth_log = ''.dup
-      proxy = WEBrick::HTTPProxyServer.new({
-        :ServerType => Thread,
-        :Logger => proxy_logger,
-        :AccessLog => [[NullLog, ""]],
-        :ProxyAuthProc => lambda {|req, res|
-          proxy_auth_log << req.request_line
-          if req["Proxy-Authorization"] != "Basic #{['user:pass'].pack('m').chomp}"
-            raise WEBrick::HTTPStatus::ProxyAuthenticationRequired
-          end
-        },
-        :BindAddress => '127.0.0.1',
-        :Port => 0})
-      _, proxy_port, _, proxy_host = proxy.listeners[0].addr
-      proxy_url = "http://#{proxy_host}:#{proxy_port}/"
-      begin
-        th = proxy.start
-        srv.mount_proc("/proxy", lambda { |req, res| res.body = "proxy" } )
-        URI.open("#{url}/proxy",
-            :proxy_http_basic_authentication=>[proxy_url, "user", "pass"]) {|f|
-          assert_equal("200", f.status[0])
-          assert_equal("proxy", f.read)
-        }
-        assert_match(/#{Regexp.quote url}/, proxy_auth_log); proxy_auth_log.clear
-        assert_raise(ArgumentError) {
-          URI.open("#{url}/proxy",
-              :proxy_http_basic_authentication=>[true, "user", "pass"]) {}
-        }
-        assert_equal("", proxy_auth_log); proxy_auth_log.clear
-      ensure
-        proxy.shutdown
-        th.join
-      end
-      assert_equal("", proxy_log.string)
-    }
-  end
-
-  def test_authenticated_proxy_http_basic_authentication_success
-    with_http {|srv, dr, url|
-      proxy_log = StringIO.new(''.dup)
-      proxy_logger = WEBrick::Log.new(proxy_log, WEBrick::BasicLog::WARN)
-      proxy_auth_log = ''.dup
-      proxy = WEBrick::HTTPProxyServer.new({
-        :ServerType => Thread,
-        :Logger => proxy_logger,
-        :AccessLog => [[NullLog, ""]],
-        :ProxyAuthProc => lambda {|req, res|
-          proxy_auth_log << req.request_line
-          if req["Proxy-Authorization"] != "Basic #{['user:pass'].pack('m').chomp}"
-            raise WEBrick::HTTPStatus::ProxyAuthenticationRequired
-          end
-        },
-        :BindAddress => '127.0.0.1',
-        :Port => 0})
-      _, proxy_port, _, proxy_host = proxy.listeners[0].addr
-      proxy_url = "http://user:pass@#{proxy_host}:#{proxy_port}/"
-      begin
-        th = proxy.start
-        srv.mount_proc("/proxy", lambda { |req, res| res.body = "proxy" } )
-        URI.open("#{url}/proxy", :proxy => proxy_url) {|f|
-          assert_equal("200", f.status[0])
-          assert_equal("proxy", f.read)
-        }
-        assert_match(/#{Regexp.quote url}/, proxy_auth_log); proxy_auth_log.clear
-        assert_equal("", proxy_auth_log); proxy_auth_log.clear
-      ensure
-        proxy.shutdown
-        th.join
-      end
-      assert_equal("", proxy_log.string)
-    }
-  end
-
   def test_redirect
     with_http {|srv, dr, url|
       srv.mount_proc("/r1/") {|req, res| res.status = 301; res["location"] = "#{url}/r2"; res.body = "r1" }
